move: useInfiniteFetchProducts, useProducts hooks products 폴더 하위로 이동
